refactor: Выполнить разбивку на  под модуль ДТО и подмудуль Сервера статистики согласно ТЗ.
