--- stats-service/src/main/java/ru/yandex/statsservice/StatsServiceApplication.java
@@ -1,13 +1,0 @@
-package ru.yandex.statsservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StatsServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(StatsServiceApplication.class, args);
-    }
-
-}
