--- stats-service/src/main/java/ru/yandex/statsservice/dto/Stat.java
@@ -1,14 +1,0 @@
-package ru.yandex.statsservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-@AllArgsConstructor
-public class Stat {
-    private final String app;
-    private final String uri;
-    private final Integer hits;
-}
