--- stats-service/src/test/java/ru/yandex/statsservice/StatsServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package ru.yandex.statsservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class StatsServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
