+++ stats/stats-dto/src/main/java/ru/practicum/statsdto/dto/Stat.java
@@ -1,0 +1,14 @@
+package ru.practicum.statsdto.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+@AllArgsConstructor
+public class Stat {
+    private final String app;
+    private final String uri;
+    private final Integer hits;
+}
