+++ stats/stats-service/src/main/java/ru/practicum/statsservice/StatsServiceApplication.java
@@ -1,0 +1,13 @@
+package ru.practicum.statsservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StatsServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(StatsServiceApplication.class, args);
+    }
+
+}
