+++ stats/stats-service/src/main/java/ru/practicum/statsservice/util/HitMapper.java
@@ -1,0 +1,19 @@
+package ru.practicum.statsservice.util;
+
+import lombok.experimental.UtilityClass;
+import ru.practicum.statsdto.dto.HitDto;
+import ru.practicum.statsservice.model.Hit;
+
+@UtilityClass
+public class HitMapper {
+    public static Hit toHit(HitDto dto) {
+        Hit hit = new Hit();
+
+        hit.setApp(dto.getApp());
+        hit.setUri(dto.getUri());
+        hit.setIp(dto.getIp());
+        hit.setTimeStamp(dto.getTimeStamp());
+
+        return hit;
+    }
+}
