+++ stats/stats-service/src/main/java/ru/practicum/statsservice/util/exception/InvalidPeriodException.java
@@ -1,0 +1,7 @@
+package ru.practicum.statsservice.util.exception;
+
+public class InvalidPeriodException extends RuntimeException {
+    public InvalidPeriodException(String message) {
+        super(message);
+    }
+}
