+++ stats/stats-service/src/main/resources/application.properties
