+++ stats/stats-service/src/main/resources/schema.sql
