+++ stats/stats-service/src/test/java/ru/practicum/statsservice/StatsServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package ru.practicum.statsservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class StatsServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
